Implement ProjectClaim objects
When calculating usage of a resource against a project, there is some
common information that we need to know as an enforcement library.
This can be referred to as a "claim", which consists of the name of
the resource being claimed, the project claiming the resources in the
request, and the quantity being claimed.

This commit adds some basic plumbing and tests that introduce this
object. Subsequent patches will introduce a context manager that
performs usage calculation using ProjecClaim objects.

Change-Id: Iacfa39535c58a7c6b49ba01c5ebf6bbb5f3fe8d1