Remove ProjectClaim object from oslo.limit
The concept of having project claims and tacking them on to an
Enforcer object is too heavy-handed for what we need right now.
Instead, we can just use a dictionary with resource names as the key
and their respective usage/request limits as values.

Future patches will remove the context manager functionality from
oslo.limit, too. We can revisit the concept of a context manager to
implement additional functionality for verification and race
conditions when we have a for sure need for it.

The goal of this patch is to make the interface as simple as possible
so that it is easier for projects to adopt it.

Note: we're able to make these changes because we haven't released
oslo.limit, yet. We can make breaking changes up to version 1.0, so we
can easily come back to this approach up to that point.

Change-Id: I3f04ffc6d27208500b8e6409c3c30e6cb3147169