Remove __enter__ and __exit__ methods from Enforcer
Originally, we wanted oslo.limit's Enforcer object to be a context
manager to make it easier to abstract logic to protect against
resource race conditions. Since verification logic isn't going to be
implemented right away, we can just focus on a single method for
enforcement instead, making it easier to adopt.

Subsequent patches are going to introduce a new enforcement interface
for oslo.limit.

Change-Id: I3a1e58cb18a498bca67f36c76d8e8c8c4fb80083