Add devstack job to .zuul.conf
We know we want to have some level of functional testing to verify
enforcement models. This commit uses one approach to add a base
devstack job to oslo.limit's .zuul.conf so that we have a minimal-ish
installation to configure with limits and write actual integration and
functional tests against.

Change-Id: I50c41805b8e196fdaf20c95e9aa687b98c369a57