Call Keystone API once to get (registered_)limits
Currently we are calling Keystone API N times where N is the number of
resources to check. We can instead call once and filter the results
by resource names if needed.

This also adds the ability for callers to get a list of limits for all
resources for a given service_id and region_id, if desired.

Change-Id: Ic4857f5553527795a4de2154aa14f93e3c0b6dc7